pyproject.toml
[project]
name = "clinical-genomics-private"
version = "0.1.0"
description = "Private clinical genomics AI benchmark and testing suite"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.20.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "pytest>=7.4.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

